\name{hyp}
\alias{hyp}
\non_function{} 
\title{Example data for MOC.}
\usage{data(hyp)}
   
\description{

  Simulated data from a model fitted on a real data set giving mother's
  report of their child hyperactivity on a Likert scale at 17,30 and 42
  months of age.
  
}

\format{
  This data frame contains 503 rows with the following columns:
  \describe{
    \item{gender}{boys and girls coded 1 and 2}
    \item{hyp1}{hyperactivity rating at 17 months}
    \item{hyp2}{hyperactivity rating at 30 months}
    \item{hyp3}{hyperactivity rating at 42 months}
  }
}


\author{Bernard Boulerice <\email{Bernard.Boulerice@umontreal.ca}>}


\keyword{datasets}
