\name{moc.dat}
\alias{moc.dat}
\non_function{} 
\title{Example data for MOC.}
\usage{data(moc)}
   
\description{

  Simulated data from a mixture model.
  
}

\format{
  This data frame contains 500 rows (subjects) with the following columns
  (variables) :
  \describe{
    \item{Y1, Y2, Y3, Y4}{simulated observation at 4 occasions.}
    \item{X1}{a dichotomous covariate.}
  }
}


\author{Bernard Boulerice <\email{Bernard.Boulerice@umontreal.ca}>}


\keyword{datasets}
