\name{post}
\alias{post}
\title{Generic function post.}

\description{

  \code{post} is a generic function to compute the posterior
  probabilities of a fitted model.
  
}

\usage{

       post(object,...)

}

\arguments{

  \item{object}{ an object for which posterior probabilities are meaningful.}

  \item{...}{ other arguments.}

}

\value{ Posterior probabilities computed from the object `object'.}

\seealso{\code{\link{post.moc}}}

\keyword{methods}
